+++ index.js
@@ -1,13 +1,16 @@
 #!/usr/bin/env node
 
 const chalk = require("chalk");
-const { question } = require("readline-sync");
+const { input } = require("@inquirer/prompts");
 
 const { GLOBAL_NAME, SHORT_NAME, VERSION, BUILD } = require("./src/variables/constants");
 
 const _intCmds = require("./src/functions/interpret");
 const _detectArgs = require("./src/functions/detectArgs");
 const _caseSensitivePath = require("./src/functions/caseSensitivePath");
+const _fatalError = require("./src/functions/fatalError");
+
+const exit = require("./src/commands/exit");
 
 const Verbose = require("./src/classes/Verbose");
 
@@ -57,14 +60,26 @@ const Verbose = require("./src/classes/Verbose");
 
   Verbose.custom("Starting command interpreter...");
   while (true) {
-    // Ask the user for a command
-    const command = question(
-      `${chalk.bold.green(SHORT_NAME.toLowerCase())} ${chalk.blueBright(
-        _caseSensitivePath(process.cwd())
-      )} ${chalk.red("$")} `
-    );
+    try {
+      const command = await input({
+        message: `${chalk.blueBright(_caseSensitivePath(process.cwd()))} ${chalk.red("$")}`,
+        theme: {
+          prefix: chalk.bold.green(SHORT_NAME.toLowerCase()),
+          style: {
+            answer: (text) => chalk.reset(text),
+          },
+        },
+      });
 
-    Verbose.custom("Interpreting command...");
-    await _intCmds(command?.trim());
+      Verbose.custom("Interpreting command...");
+      await _intCmds(command?.trim());
+    } catch ({ name }) {
+      if (name === "ExitPromptError") {
+        // If the user presses Ctrl+C, exit BubbleOS gracefully
+        // Note this does not catch if a command is executing
+        Verbose.custom("Detected Ctrl+C, exiting...");
+        exit();
+      }
+    }
   }
 })();